Day 25: Running Time and Complexity
Task:
A prime is a natural number greater than 1 that has no positive divisors other than 1 and itself. Given a number, n , determine and print whether it's Prime or Not prime.

Input Format:
The first line contains an integer, T , the number of test cases.
Each of the T subsequent lines contains an integer,n , to be tested for primality.