Day 29: Bitwise AND
Objective:
Welcome to the last day! Today, we're discussing bitwise operations. Check out the Tutorial tab for learning materials and an instructional video!

Task:
Given set S = {1,2,3, ... , N}. Find two integers, A and B (where A < B), from set S such that the value of  is the maximum possible and also less than a given integer, K. In this case,&  represents the bitwise AND operator.

Input Format:
The first line contains an integer,T , the number of test cases.
Each of the T subsequent lines defines a test case as 2 space-separated integers, N and K, respectively.

Constraints:
*1<=T<=10*3
*1<=T<=10*3

Output Format:
For each test case, print the maximum possible value of A&B on a new line.